+++ src/views/TreasuryDashboard/components/Graph/TreasuryAssetsTable.tsx
@@ -4,7 +4,7 @@ import { useMemo, useState } from "react";
 import { TokenRecord_Filter, TokenRecordsDocument } from "src/generated/graphql";
 import { formatCurrency, formatNumber } from "src/helpers";
 import { renameToken } from "src/helpers/subgraph/ProtocolMetricsHelper";
-import { useTokenRecordsQueries } from "src/hooks/useTokenRecords";
+import { useTokenRecordsQueries } from "src/hooks/useSubgraphTokenRecords";
 import { ChartCard } from "src/views/TreasuryDashboard/components/Graph/ChartCard";
 import {
   AssetsTableProps,
